# Conflicts:
#	wadsrc/static/zscript/menu/optionmenuitems.txt

+++ src/gl/models/gl_models.cpp
@@ -157,7 +157,7 @@ void FGLModelRenderer::DrawElements(int numIndices, size_t offset)
 
 double FGLModelRenderer::GetTimeFloat()
 {
-	return (float)I_msTime() * (float)TICRATE / 1000.0f;
+	return (double)I_msTime() * (double)TICRATE / 1000.;
 }
 
 //===========================================================================
